refactor(server): extract announce scheduler into AutoAnnounceService (#3962 4.2a PR3 §4b)
Extracts startAnnounceScheduler/setAnnounceInterval/restartAnnounceScheduler/
sendAutoAnnouncement/previewAnnouncementMessage + the announceScheduler field
out of MeshtasticManager into services/autoAnnounceService.ts, following the
constructor-injected `import type` pattern from nodeDbMaintenanceService
(PR2). replaceAnnouncementTokens deliberately stays on the manager (shared
with acknowledgement/welcome/geofence/timer token expansion) but its
visibility widens from private to public so the extracted service can call
back into it via the injected `mgr` reference. Two narrow public accessors
(isRebootMergeInProgress, and reuse of the existing isDeviceConnected) bridge
the remaining private state the service needs, without widening those fields.

Also fixes the disconnect-gap (#3962 4.2a spec §5b): disconnect() stopped
every other periodic scheduler (traceroute/remoteLocalStats/
remoteAdminScanner/timeSync/distanceDelete/localStats/timeOffset) but never
announceScheduler — only userDisconnect() did. An unexpected disconnect left
the announce scheduler ticking (self-guarded, so harmless-but-wrong) instead
of being torn down like its siblings. disconnect() now calls
autoAnnounceService.stop() alongside the other scheduler stops;
handleConnected's existing startAnnounceScheduler() call re-arms it on
(re)connect, so nothing regresses for the normal reconnect path.

Manager thin delegates preserve the existing public API (sendAutoAnnouncement/
previewAnnouncementMessage/restartAnnounceScheduler/setAnnounceInterval) for
announceRoutes.ts and server.ts's per-source callback wiring.
