+++ src/server/meshtasticManager.ts
@@ -62,7 +62,6 @@ import { normalizeChannelRole } from './constants/channelRole.js';
 import { isAutoFavoriteEligible } from './constants/autoFavorite.js';
 import { createRequire } from 'module';
 import { validateCron, scheduleCron, type CronJob } from './utils/cronScheduler.js';
-import { CronOrIntervalScheduler, type ScheduleMode } from './services/cronOrIntervalScheduler.js';
 import {
   shouldExcludeFromPacketLog,
   isPhantomInternalPacket,
@@ -71,6 +70,7 @@ import {
   matchesMqttEcho,
 } from './services/mqttProxyBridge.js';
 import { NodeDbMaintenanceService } from './services/nodeDbMaintenanceService.js';
+import { AutoAnnounceService } from './services/autoAnnounceService.js';
 import fs from 'fs';
 import path from 'path';
 import * as net from 'net';
@@ -543,7 +543,6 @@ class MeshtasticManager implements ISourceManager {
   private timeOffsetSamples: number[] = [];
   private timeOffsetInterval: NodeJS.Timeout | null = null;
   private localStatsIntervalMinutes: number = 15;  // Default 5 minutes
-  private announceScheduler: CronOrIntervalScheduler | null = null;
   private timerCronJobs: Map<string, CronJob> = new Map();
   private geofenceNodeState: Map<string, Set<number>> = new Map(); // geofenceId -> set of nodeNums currently inside
   private geofenceWhileInsideTimers: Map<string, NodeJS.Timeout> = new Map(); // geofenceId -> interval timer
@@ -921,9 +920,14 @@ class MeshtasticManager implements ISourceManager {
   // constructor (import-cycle-safe: the service only `import type`s MeshtasticManager).
   private readonly nodeDbMaintenanceService: NodeDbMaintenanceService;
 
+  // Auto-announce scheduler + send — extracted to a service (#3962 Phase 4.2a
+  // PR3 §4b). Same injection pattern as nodeDbMaintenanceService above.
+  private readonly autoAnnounceService: AutoAnnounceService;
+
   constructor(sourceId: string = 'default', sourceConfig?: { host?: string; port?: number; heartbeatIntervalSeconds?: number; virtualNode?: VirtualNodeConfig; mqttLink?: MeshtasticMqttLink; passiveMode?: boolean; passiveResyncStaleMs?: number | null }) {
     this.sourceId = sourceId;
     this.nodeDbMaintenanceService = new NodeDbMaintenanceService(this);
+    this.autoAnnounceService = new AutoAnnounceService(this);
     if (sourceConfig) {
       this.sourceConfigOverride = {
         host: sourceConfig.host,
@@ -1866,6 +1870,13 @@ class MeshtasticManager implements ISourceManager {
 
     // Stop auto-delete-by-distance scheduler
     this.stopDistanceDeleteScheduler();
+
+    // Stop announce scheduler (#3962 Phase 4.2a PR3 §5b — disconnect() used to
+    // omit this; only userDisconnect() stopped it, so an unexpected disconnect
+    // left the scheduler ticking, self-guarded but latent. handleConnected()
+    // re-arms it via startAnnounceScheduler() on (re)connect, so stopping here
+    // is safe.)
+    this.autoAnnounceService.stop();
 
     // Stop LocalStats collection
     this.stopLocalStatsScheduler();
@@ -2895,120 +2906,22 @@ class MeshtasticManager implements ISourceManager {
     }
   }
 
+  /**
+   * Thin delegate — arming logic lives in `AutoAnnounceService`
+   * (#3962 Phase 4.2a PR3 §4b).
+   */
   private async startAnnounceScheduler(): Promise<void> {
-    // Clear any existing scheduler
-    this.announceScheduler?.stop();
-    this.announceScheduler = null;
-
-    // Check if auto-announce is enabled
-    const autoAnnounceEnabled = await databaseService.settings.getSettingForSource(this.sourceId, 'autoAnnounceEnabled');
-    if (autoAnnounceEnabled !== 'true') {
-      logger.debug('📢 Auto-announce is disabled');
-      return;
-    }
-
-    // Determine schedule mode (cron vs interval — per-source, written by
-    // AutoAnnounceSection via /api/settings?sourceId=)
-    const useSchedule = (await databaseService.settings.getSettingForSource(this.sourceId, 'autoAnnounceUseSchedule')) === 'true';
-
-    let mode: ScheduleMode;
-    if (useSchedule) {
-      const scheduleExpression = (await databaseService.settings.getSettingForSource(this.sourceId, 'autoAnnounceSchedule')) || '0 */6 * * *';
-      logger.debug(`📢 Starting announce scheduler with cron expression: ${scheduleExpression}`);
-      mode = { kind: 'cron', expression: scheduleExpression };
-    } else {
-      const intervalHours = parseInt((await databaseService.settings.getSettingForSource(this.sourceId, 'autoAnnounceIntervalHours')) || '6');
-      const intervalMs = intervalHours * 60 * 60 * 1000;
-      logger.debug(`📢 Starting announce scheduler with ${intervalHours} hour interval`);
-      mode = { kind: 'interval', intervalMs };
-    }
-
-    this.announceScheduler = new CronOrIntervalScheduler({
-      label: `Meshtastic:${this.sourceId}`,
-      mode,
-      onTick: () => {
-        logger.debug(`📢 Announce tick triggered (connected: ${this.isConnected})`);
-        if (this.isConnected) {
-          return this.sendAutoAnnouncement(true).catch((error: Error) => {
-            logger.error('❌ Error in auto-announce:', error);
-          });
-        }
-        logger.debug('📢 Skipping announcement - not connected to node');
-      },
-    });
-
-    if (!this.announceScheduler.start()) {
-      // cron expression was invalid; warning already logged by the scheduler
-      if (mode.kind === 'cron') {
-        logger.error(`❌ Invalid cron expression: ${mode.expression}`);
-      }
-      this.announceScheduler = null;
-      return;
-    }
-
-    logger.debug('📢 Announce scheduler started');
-
-    // Check if announce-on-start is enabled (per-source; applies to both cron and interval modes)
-    const announceOnStart = await databaseService.settings.getSettingForSource(this.sourceId, 'autoAnnounceOnStart');
-    if (announceOnStart === 'true') {
-      // Check spam protection: don't send if announced within last hour
-      const lastAnnouncementTime = await databaseService.settings.getSettingForSource(this.sourceId, 'lastAnnouncementTime');
-      const now = Date.now();
-      const oneHour = 60 * 60 * 1000;
-
-      if (lastAnnouncementTime) {
-        const timeSinceLastAnnouncement = now - parseInt(lastAnnouncementTime);
-        if (timeSinceLastAnnouncement < oneHour) {
-          const minutesRemaining = Math.ceil((oneHour - timeSinceLastAnnouncement) / 60000);
-          logger.debug(`📢 Skipping startup announcement - last announcement was ${Math.floor(timeSinceLastAnnouncement / 60000)} minutes ago (spam protection: ${minutesRemaining} minutes remaining)`);
-        } else {
-          logger.debug('📢 Sending startup announcement');
-          // Delay startup announcement to allow reboot detection and ghost cleanup to complete
-          setTimeout(async () => {
-            if (this.isConnected) {
-              try {
-                await this.sendAutoAnnouncement(true);
-              } catch (error) {
-                logger.error('❌ Error in startup announcement:', error);
-              }
-            }
-          }, 30000);
-        }
-      } else {
-        // No previous announcement, send one
-        logger.debug('📢 Sending first startup announcement');
-        // Delay startup announcement to allow reboot detection and ghost cleanup to complete
-        setTimeout(async () => {
-          if (this.isConnected) {
-            try {
-              await this.sendAutoAnnouncement(true);
-            } catch (error) {
-              logger.error('❌ Error in startup announcement:', error);
-            }
-          }
-        }, 30000);
-      }
-    }
+    return this.autoAnnounceService.startAnnounceScheduler();
   }
 
+  /** Thin delegate — see `AutoAnnounceService.setAnnounceInterval`. */
   setAnnounceInterval(hours: number): void {
-    if (hours < 3 || hours > 24) {
-      throw new Error('Announce interval must be between 3 and 24 hours');
-    }
-
-    logger.debug(`📢 Announce interval updated to ${hours} hours`);
-
-    if (this.isConnected) {
-      this.startAnnounceScheduler().catch(err => logger.error('Error starting announce scheduler:', err));
-    }
+    this.autoAnnounceService.setAnnounceInterval(hours);
   }
 
+  /** Thin delegate — see `AutoAnnounceService.restartAnnounceScheduler`. */
   restartAnnounceScheduler(): void {
-    logger.debug('📢 Restarting announce scheduler due to settings change');
-
-    if (this.isConnected) {
-      this.startAnnounceScheduler().catch(err => logger.error('Error restarting announce scheduler:', err));
-    }
+    this.autoAnnounceService.restartAnnounceScheduler();
   }
 
   /**
@@ -11668,98 +11581,9 @@ class MeshtasticManager implements ISourceManager {
     return result;
   }
 
+  /** Thin delegate — see `AutoAnnounceService.sendAutoAnnouncement`. */
   async sendAutoAnnouncement(triggeredByAutomation = false): Promise<void> {
-    if (this.rebootMergeInProgress) {
-      logger.debug('📢 Skipping auto-announcement - reboot merge in progress');
-      return;
-    }
-
-    // Airtime cutoff: skip scheduled announcements while the mesh is congested.
-    // Manual "Send Announcement" requests (triggeredByAutomation = false) are
-    // always allowed through.
-    if (triggeredByAutomation && await this.isAutomationAirtimeGated()) {
-      return;
-    }
-
-    try {
-      // All auto-announce settings are per-source (written by AutoAnnounceSection
-      // via /api/settings?sourceId=).
-      const settings = databaseService.settings;
-      const sourceId = this.sourceId;
-
-      const message = await settings.getSettingForSource(sourceId, 'autoAnnounceMessage') || 'MeshMonitor {VERSION} online for {DURATION} {FEATURES}';
-
-      // Multi-channel support: read JSON array, fall back to legacy single index
-      let channelIndexes: number[];
-      const channelIndexesStr = await settings.getSettingForSource(sourceId, 'autoAnnounceChannelIndexes');
-      if (channelIndexesStr) {
-        try {
-          const parsed = JSON.parse(channelIndexesStr);
-          channelIndexes = Array.isArray(parsed) ? parsed.filter(n => typeof n === 'number') : [0];
-        } catch {
-          channelIndexes = [0];
-        }
-      } else {
-        // Legacy migration: read old single channel setting (pre-4.0, global-only)
-        const legacyIndex = parseInt(await settings.getSetting('autoAnnounceChannelIndex') || '0');
-        channelIndexes = [legacyIndex];
-      }
-
-      if (channelIndexes.length === 0) {
-        channelIndexes = [0];
-      }
-
-      // Replace tokens
-      const replacedMessage = await this.replaceAnnouncementTokens(message);
-
-      logger.debug(`📢 Sending auto-announcement to ${channelIndexes.length} channel(s) [${channelIndexes.join(',')}]: "${replacedMessage}"`);
-
-      channelIndexes.forEach((channelIdx, i) => {
-        this.messageQueue.enqueue(
-          replacedMessage,
-          0, // destination: 0 for channel broadcast
-          undefined, // no reply-to for announcements
-          () => {
-            logger.debug(`\u2705 Auto-announcement ${i + 1}/${channelIndexes.length} delivered to channel ${channelIdx}`);
-          },
-          (reason: string) => {
-            logger.warn(`\u274c Auto-announcement ${i + 1}/${channelIndexes.length} failed on channel ${channelIdx}: ${reason}`);
-          },
-          channelIdx, // channel number
-          1 // single attempt, no retry for broadcasts
-        );
-      });
-
-      // Update last announcement time (per-source)
-      if (this.sourceId) {
-        await databaseService.settings.setSourceSetting(this.sourceId, 'lastAnnouncementTime', Date.now().toString());
-      } else {
-        await databaseService.settings.setSetting('lastAnnouncementTime', Date.now().toString());
-      }
-      logger.debug('📢 Last announcement time updated');
-
-      // Check if NodeInfo broadcasting is enabled (per-source)
-      const nodeInfoEnabled = await settings.getSettingForSource(sourceId, 'autoAnnounceNodeInfoEnabled') === 'true';
-      if (nodeInfoEnabled) {
-        try {
-          const nodeInfoChannelsStr = await settings.getSettingForSource(sourceId, 'autoAnnounceNodeInfoChannels') || '[]';
-          const nodeInfoChannels = JSON.parse(nodeInfoChannelsStr) as number[];
-          const nodeInfoDelaySeconds = parseInt(await settings.getSettingForSource(sourceId, 'autoAnnounceNodeInfoDelaySeconds') || '30');
-
-          if (nodeInfoChannels.length > 0) {
-            logger.debug(`📢 NodeInfo broadcasting enabled - will broadcast to ${nodeInfoChannels.length} channel(s)`);
-            // Run NodeInfo broadcasting asynchronously (don't block the announcement)
-            this.broadcastNodeInfoToChannels(nodeInfoChannels, nodeInfoDelaySeconds).catch(error => {
-              logger.error('❌ Error in NodeInfo broadcasting:', error);
-            });
-          }
-        } catch (parseError) {
-          logger.error('❌ Error parsing NodeInfo channels setting:', parseError);
-        }
-      }
-    } catch (error) {
-      logger.error('❌ Error sending auto-announcement:', error);
-    }
+    return this.autoAnnounceService.sendAutoAnnouncement(triggeredByAutomation);
   }
 
   /**
@@ -11795,7 +11619,14 @@ class MeshtasticManager implements ISourceManager {
     return args;
   }
 
-  private async replaceAnnouncementTokens(message: string, urlEncode: boolean = false): Promise<string> {
+  /**
+   * Visibility widened from `private` to (default) `public` so `AutoAnnounceService`
+   * (#3962 Phase 4.2a PR3 §4b) can call back into it via the injected `mgr`
+   * reference. Body/location unmoved — still shared by
+   * `replaceAcknowledgementTokens`/`replaceWelcomeTokens`/`replaceGeofenceTokens`
+   * and the auto-responder's timer text path, all of which call it internally.
+   */
+  async replaceAnnouncementTokens(message: string, urlEncode: boolean = false): Promise<string> {
     // Defensive coercion: callers come from settings/DB and protobuf paths where the static type
     // is `string` but the runtime value isn't always proven to be one. CodeQL flagged every
     // `result.includes('{TOKEN}')` below as type-confusion-through-parameter-tampering without this.
@@ -11966,10 +11797,11 @@ class MeshtasticManager implements ISourceManager {
   }
 
   /**
-   * Public wrapper for replaceAnnouncementTokens, used by the preview API endpoint.
+   * Thin delegate — see `AutoAnnounceService.previewAnnouncementMessage`.
+   * Used by the preview API endpoint.
    */
   public async previewAnnouncementMessage(message: string): Promise<string> {
-    return this.replaceAnnouncementTokens(message);
+    return this.autoAnnounceService.previewAnnouncementMessage(message);
   }
 
   /**
@@ -14193,6 +14025,14 @@ class MeshtasticManager implements ISourceManager {
     this.deviceNodeNums.delete(nodeNum);
   }
 
+  // ── Narrow accessor for AutoAnnounceService (#3962 Phase 4.2a PR3 §4b) ──
+  // Bridges the private `rebootMergeInProgress` guard without widening the
+  // field itself — same rationale as the NodeDbMaintenanceService block above.
+  /** `true` while a post-reboot NodeDB merge is suppressing broadcasts. */
+  isRebootMergeInProgress(): boolean {
+    return this.rebootMergeInProgress;
+  }
+
   // Async version that fetches uptimes in a single bulk query - works with all DB backends
   async getAllNodesAsync(sourceId?: string): Promise<DeviceInfo[]> {
     return this.nodeDbMaintenanceService.getAllNodesAsync(sourceId);
@@ -14433,12 +14273,8 @@ class MeshtasticManager implements ISourceManager {
       this.timeSyncInterval = null;
     }
 
-    // Stop announce scheduler if active
-    if (this.announceScheduler) {
-      this.announceScheduler.stop();
-      this.announceScheduler = null;
-      logger.debug('📢 Stopped announce scheduler');
-    }
+    // Stop announce scheduler if active (idempotent — no-op if not armed)
+    this.autoAnnounceService.stop();
 
     // Stop all timer cron jobs
     this.timerCronJobs.forEach((job, id) => {
